+++ pages/api/cron/weekly.ts
@@ -1,4 +1,4 @@
-import sendScheduledEmails from "cron/scheduledEmails";
+import sendWeeklyProgressReport from "cron/weeklyProgressReport";
 import { NextApiRequest, NextApiResponse } from "next";
 
 const CRON_SECRET_SIGNING_KEY = process.env.CRON_SECRET_SIGNING_KEY;
@@ -8,7 +8,7 @@ if (!CRON_SECRET_SIGNING_KEY)
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.headers["signing-key"] === CRON_SECRET_SIGNING_KEY) {
     try {
-      await sendScheduledEmails();
+      await sendWeeklyProgressReport();
       res.status(200).json({ success: true });
     } catch (err) {
       res.status(500).json({ success: false, message: (err as any).message });