+++ src/pages/HomePage/Heatmap.tsx
@@ -72,7 +72,7 @@ export default function Heatmap({ history }: { history: HistorySegment[] }) {
           tooltipDataAttrs={(value: HistoryVal) => {
             if (!value?.date) return;
             return {
-              "data-tip": genDataTip(value),
+              "data-tip": `${genDataTip(value)} on ${new Date(value.date).toDateString().slice(0, 10)}`,
             };
           }}
         />